+++ R/classes.R
@@ -323,7 +323,7 @@ setClass(
 #' @name processParam
 #' @description
 #' Utility class that defines a data processing procedure and any params used
-#' in performing it. Packages defining processing methods will create their own 
+#' in performing it. Packages defining processing methods will create their own
 #' child classes. These parameter objects are intended to be passed alongside
 #' the data to process to [processData()].
 #' @slot param list. Named parameters to use with the intended processing
@@ -339,7 +339,7 @@ setClass("processParam", contains = "VIRTUAL", slots = list(param = "list"))
 #' @title Spatial Value Key
 #' @description
 #' A metaprogramming object that references a set of information to get
-#' from a `giotto` object when used as `svkey@get(gobject)`. 
+#' from a `giotto` object when used as `svkey@get(gobject)`.
 #' Referenced data will be retrieved as a `data.table` via [spatValues()]
 #' @keywords internal
 setClass("svkey",
@@ -1387,7 +1387,7 @@ giottoPolygon <- setClass(
 #' @export
 updateGiottoPolygonObject <- function(gpoly) {
     if (!inherits(gpoly, "giottoPolygon")) {
-        stop("This function is only for giottoPoints")
+        stop("This function is only for giottoPolygon")
     }
 
     # 3.2.X adds cacheing of IDs
@@ -1396,6 +1396,10 @@ updateGiottoPolygonObject <- function(gpoly) {
             as.list(gpoly@spatVector)$poly_ID
         )
     }
+
+    # 0.4.7 changes overlaps representation
+    # intersection `SpatVector` -> `overlapInfo`-inheriting classes
+    gpoly <- .update_overlaps(gpoly)
 
     gpoly
 }
@@ -1516,10 +1520,120 @@ setClass(
 
 
 
+## overlapInfo ####
+setClass("overlapInfo",
+    contains = c("spatFeatData", "VIRTUAL"),
+    slots = list(data = "ANY")
+)
+setClass("overlapPoint", contains = c("overlapInfo", "VIRTUAL"))
+setClass("overlapIntensity", contains = c("overlapInfo", "VIRTUAL"))
+
+setClass("overlapPointDT",
+    contains = "overlapPoint",
+    slots = list(
+        spat_ids = "character", # spat_ids are unique, no need to record npoly
+        feat_ids = "character",
+        nfeats = "integer"
+    )
+)
+setClass("overlapIntensityDT",
+    contains = "overlapIntensity",
+    slots = list(
+        nfeats = "integer", # skip spat_ids/feat_ids. This data is not sparse
+        fun = "character"
+    )
+)
 
 
 
+# update old overlaps information to new `overlapInfo`
+.update_overlaps <- function(x, ...) {
 
+    if (inherits(x, "giottoPolygon")) {
+        res <- .update_overlaps(x@overlaps,
+            poly_ids = x$poly_ID,
+            spat_unit = spatUnit(x),
+            ...
+        )
+        x@overlaps <- res
+        return(x)
+    }
+    if (inherits(x, "list")) {
+        list_names <- names(x)
+        list_res <- lapply(list_names, function(ovlp_name) {
+            if (ovlp_name == "intensity") {
+                # recurse over list of intensity overlaps (can't set feat_type)
+                intensity_res <- .update_overlaps(x[["intensity"]], ...)
+                return(intensity_res)
+            } else {
+                updated_res <- .update_overlaps(x[[ovlp_name]],
+                    feat_type = ovlp_name,
+                    ...
+                )
+                return(updated_res)
+            }
+        })
+        names(list_res) <- list_names
+        return(list_res)
+    }
+    if (inherits(x, "SpatVector")) {
+        return(.update_overlaps_points(x, ...))
+    }
+    if (inherits(x, "data.table")) {
+        return(.update_overlaps_intensity(x, ...))
+    }
+    x # allow passthrough if not matching either signature
+}
+
+#' @param x (SpatVector) the old overlaps representation to convert
+#' @param poly_ids the `spatIDs()` of the giottoPolygon. This is since the
+#' ordering of the polygon IDs within the overlaps data usually does not match.
+#' @param spat_unit,feat_type spat_unit / feat_type
+#' @noRd
+.update_overlaps_points <- function(x, poly_ids,
+    spat_unit = NA_character_, feat_type = NA_character_, ...) {
+    checkmate::assert_character(poly_ids)
+    data <- terra::as.data.frame(x)
+    data.table::setDT(data)
+    odt <- new("overlapPointDT",
+        spat_unit = spat_unit,
+        feat_type = feat_type,
+        provenance = spat_unit,
+        nfeats = as.integer(nrow(data))
+    )
+    odt@feat_ids <- unique(data$feat_ID)
+    odt@spat_ids <- unique(poly_ids)
+    data[, feat := as.integer(feat_ID_uniq)]
+    data[, feat_ID_uniq := NULL]
+    data.table::setnames(data,
+        old = c("poly_ID", "feat_ID"),
+        new = c("poly", "feat_id_index")
+    )
+    data <- data[!is.na(poly) & !is.na(feat),]    # drop NAs
+    # Ensure data is stored as integer-based mapping
+    data[, poly := match(poly, odt@spat_ids)]
+    data[, feat_id_index := match(feat_id_index, odt@feat_ids)]
+    data.table::setkeyv(data, "feat")
+    data.table::setindexv(data, "poly")
+    data.table::setcolorder(data, c("poly", "feat", "feat_id_index"))
+    # add to object
+    odt@data <- data
+    odt
+}
+
+.update_overlaps_intensity <- function(x,
+    spat_unit = NA_character_, feat_type = NA_character_, ...) {
+    odt <- new("overlapIntensityDT",
+        spat_unit = spat_unit,
+        feat_type = feat_type,
+        provenance = spat_unit,
+        nfeats = as.integer(ncol(x) - 1)
+    )
+    fids <- setdiff(names(x), "poly_ID")
+    x <- x[, lapply(.SD, sum), by = "poly_ID", .SDcols = fids]
+    odt@data <- x
+    odt
+}
 
 ## featureNetwork class ####